moi_admin/requirements.txt
channels==3.0.4
channels_redis
Django==4.0
djangorestframework==3.13.1
face-recognition==1.3.0
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.0
psycopg2==2.9.3
